+++ qualitycheck-20190115/alibabacloud_qualitycheck20190115/models/_upload_audio_data_request.py
@@ -10,9 +10,9 @@ class UploadAudioDataRequest(DaraModel):
         base_me_agent_id: int = None,
         json_str: str = None,
     ):
-        # The workspace ID. This ID is used to select a specific workspace in a multi-workspace scenario. The default workspace is used if you do not specify this parameter.
+        # The business space ID. In multi-business space scenarios, this parameter specifies the business space to use. Default value: the default business space.
         self.base_me_agent_id = base_me_agent_id
-        # The complete JSON string. For more information, see the details below.
+        # The complete JSON string. For details about the content, see the following detailed information.
         # 
         # This parameter is required.
         self.json_str = json_str